kamu-auth-rebac: extract entities/*
